move getImageDataUrl to the shared imageDataUrl module and adopt at remaining sites
