+++ src/vs/workbench/contrib/positronNotebook/browser/notebookMimeUtils.ts
@@ -3,6 +3,12 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { isSvgMimeType } from '../../../services/positronPlots/common/imageDataUrl.js';
+
+// Re-export so existing consumers keep importing the SVG check from this
+// module; the definition lives with the image data URL encode/decode helpers.
+export { isSvgMimeType };
+
 /**
  * Image MIME types that should be base64 encoded when serializing notebook outputs.
  * Excludes SVG (image/svg+xml) which is text-based XML.
@@ -53,19 +59,6 @@ const TEXT_BASED_MIME_TYPES_SET = new Set<string>(TEXT_BASED_MIME_TYPES);
  */
 export function isImageMimeType(mimeType: string): boolean {
 	return IMAGE_MIME_TYPES_SET.has(mimeType.toLowerCase());
-}
-
-/**
- * Checks if a MIME type is SVG (image/svg+xml), ignoring MIME parameters
- * (e.g. '; charset=utf-8'). SVG is text-based XML rather than binary image
- * data, so it is classified separately from other images.
- * @param mimeType The MIME type to check.
- * @returns True if the MIME type is image/svg+xml.
- */
-export function isSvgMimeType(mimeType: string): boolean {
-	const semicolonIndex = mimeType.indexOf(';');
-	const baseMimeType = semicolonIndex >= 0 ? mimeType.substring(0, semicolonIndex) : mimeType;
-	return baseMimeType.trim().toLowerCase() === 'image/svg+xml';
 }
 
 /**
